+++ src/main/kotlin/icu/windea/pls/config/cwt/CwtConfigGroup.kt
@@ -3,7 +3,7 @@ package icu.windea.pls.config.cwt
 import com.intellij.openapi.project.*
 import icu.windea.pls.config.cwt.config.*
 import icu.windea.pls.config.cwt.config.ext.*
-import icu.windea.pls.config.cwt.config.settings.*
+import icu.windea.pls.config.cwt.setting.*
 import icu.windea.pls.core.annotations.*
 import icu.windea.pls.core.model.*
 
@@ -12,6 +12,7 @@ interface CwtConfigGroup {
 	val project: Project
 	
 	val foldingSettings: Map<String, Map<@CaseInsensitive String, CwtFoldingSetting>> //EXTENDED BY PLS
+	val postfixTemplateSettings: Map<String, Map<@CaseInsensitive String, CwtPostfixTemplateSetting>> //EXTENDED BY PLS
 	
 	val systemScopes: Map<@CaseInsensitive String, CwtSystemScopeConfig> //EXTENDED BY PLS
 	val localisationLocales: Map<String, CwtLocalisationLocaleConfig> //EXTENDED BY PLS