+++ src/main/kotlin/icu/windea/pls/config/cwt/setting/CwtFoldingSetting.kt
@@ -1,0 +1,9 @@
+package icu.windea.pls.config.cwt.setting
+
+data class CwtFoldingSetting(
+	val name: String,
+	val key: String?,
+	val keys: List<String>?,
+	val placeholder: String
+) : CwtSetting
+