+++ src/main/kotlin/icu/windea/pls/config/cwt/setting/CwtSetting.kt
@@ -1,0 +1,3 @@
+package icu.windea.pls.config.cwt.setting
+
+interface CwtSetting